Upgraded to Hibernate 3.2.1    The code changes were pulled from the fix branch and V1.4 (HIBERNATE_UPGRADE 5207 5293)    The jars were manually dropped in again due to the new directory structure
git-svn-id: https://svn.alfresco.com/repos/alfresco-enterprise/alfresco/HEAD/root@5303 c4b6b30b-aa2e-2d43-bbcb-ca4b014f7261

+++ source/java/org/alfresco/repo/node/BaseNodeServiceTest.java
@@ -406,7 +406,7 @@ public abstract class BaseNodeServiceTest extends BaseSpringTest
             .setString(1, nodeRef.getStoreRef().getProtocol())
             .setString(2, nodeRef.getStoreRef().getIdentifier())
             .list();
-        Integer count = (Integer) results.get(0);
+        Long count = (Long) results.get(0);
         return count.intValue();
     }
